+++ compiler-plugin/src/main/kotlin/tech/mappie/util/Identifiers.kt
@@ -3,9 +3,6 @@ package tech.mappie.util
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
-import tech.mappie.api.config.UseDefaultArguments
-import tech.mappie.api.config.UseStrictEnums
-import tech.mappie.api.config.UseStrictVisibility
 
 val IDENTIFIER_MAPPING = Name.identifier("mapping")
 
@@ -63,12 +60,14 @@ val CLASS_ID_OBJECT_MAPPING_CONSTRUCTOR = ClassId(PACKAGE_TECH_MAPPIE_API, Name.
 
 val CLASS_ID_ENUM_MAPPING_CONSTRUCTOR = ClassId(PACKAGE_TECH_MAPPIE_API, Name.identifier("EnumMappingConstructor"))
 
+val CLASS_ID_MAPPIE = ClassId(PACKAGE_TECH_MAPPIE_API, Name.identifier("Mappie"))
+
 val CLASS_ID_ENUM_MAPPIE = ClassId(PACKAGE_TECH_MAPPIE_API, Name.identifier("EnumMappie"))
 
 val CLASS_ID_OBJECT_MAPPIE = ClassId(PACKAGE_TECH_MAPPIE_API, Name.identifier("ObjectMappie"))
 
-val CLASS_ID_USE_STRICT_VISIBILITY = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier(UseStrictVisibility::class.simpleName!!))
+val CLASS_ID_USE_STRICT_VISIBILITY = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier("UseStrictVisibility"))
 
-val CLASS_ID_USE_DEFAULT_ARGUMENTS = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier(UseDefaultArguments::class.simpleName!!))
+val CLASS_ID_USE_DEFAULT_ARGUMENTS = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier("UseDefaultArguments"))
 
-val CLASS_ID_USE_STRICT_ENUMS = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier(UseStrictEnums::class.simpleName!!))
+val CLASS_ID_USE_STRICT_ENUMS = ClassId(PACKAGE_TECH_MAPPIE_API_CONFIG, Name.identifier("UseStrictEnums"))