+++ gradle-plugin/src/test/kotlin/tech/mappie/testing/TestBase.kt
@@ -21,7 +21,7 @@ abstract class TestBase {
 
     protected open val gradleVersion: String? = null
 
-    protected open val kotlinVersion = "2.1.0"
+    protected open val kotlinVersion = "2.1.10"
 
     @BeforeEach
     fun setup() {
@@ -74,6 +74,7 @@ abstract class TestBase {
                 id("org.jetbrains.kotlin.jvm") version "$kotlinVersion"
                 id("tech.mappie.plugin") version "$version"
             }
+
             repositories {
                 mavenLocal()
                 mavenCentral()
@@ -83,7 +84,6 @@ abstract class TestBase {
                 testImplementation(kotlin("test"))
             }
 
-            
             tasks.test {
                 useJUnitPlatform()
                 testLogging {
@@ -96,9 +96,10 @@ abstract class TestBase {
     }
 
     private fun multiplatform() {
-        directory.resolve("src/commonMain/kotlin").mkdirs()
-        directory.resolve("src/commonTest/kotlin").mkdirs()
-        directory.resolve("src/jvmMain/kotlin").mkdirs()
+        listOf("common", "jvm", "mingwX64", "ios").forEach { platform ->
+            directory.resolve("src/${platform}Main/kotlin").mkdirs()
+            directory.resolve("src/${platform}Test/kotlin").mkdirs()
+        }
 
         kotlin("build.gradle.kts",
             """
@@ -114,16 +115,16 @@ abstract class TestBase {
 
             kotlin {
                 applyDefaultHierarchyTemplate()
-                
-                sourceSets {
-                    val commonTest by getting {
-                        dependencies {
-                            implementation(kotlin("test"))
-                        }
-                    }
-                }
             
                 jvm()
+                mingwX64()
+                iosX64()
+                
+                sourceSets {
+                    commonTest.dependencies {
+                        implementation(kotlin("test"))
+                    }
+                }
             }
             """.trimIndent()
         )