+++ website/src/changelog.md
@@ -2,6 +2,10 @@
 title: "Changelog"
 layout: "layouts/changelog.html"
 changelog:
+    - date: "tbd"
+      title: "v1.1.0"
+      items:
+        - "#[161](https://github.com/Mr-Mappie/mappie/issues/161) fixed compilation on non-jvm platforms."
     - date: "2025-02-07"
       title: "v1.0.0"
       items: